LPD-41743 Swap it. BaseService has the logic, and BaseRestController extends BaseService but has the annotation. This is cleaner and just makes more sense.
